+++ src/BinaryFileHandler/ISO.h
@@ -103,7 +103,7 @@ public:
     void injectCode(const std::vector< std::pair<uint32_t, uint32_t> >&);
 
     /* calculate check sum of iso file */
-    uint64_t checkSum() const;
+    uint64_t checkSumDol() const;
 };
 
 #endif